user update test working and 1 test with multiple test groups
